See below
Player's now have a glow up tile to deselect them, players can board the ship when enough ore is deposited, the game UI is adjusted and we now have win and lose states. Winning is defined in the rocket if all three players board the ship. Losing is defined in the players if one dies